+++ nvim/init.lua
@@ -1,25 +1,2 @@
---require("user.impatient")
-require("user.options")
-require("user.keymaps")
-require("user.lsp")
---require("user.cmp")
-require("user.plugins")
-require("user.comment")
-require("user.autopairs")
-require("user.telescope")
-require("user.treesitter")
-require("user.colorscheme")
-require("user.null-ls")
-require("user.lualine")
-require("user.toggle-term")
-require("user.escape")
-require("user.navigator")
-require("user.bufferline")
-require("user.colorizer")
-require("user.context")
-require("user.nvim-tree-conf")
-require("user.firenvim-conf")
-require("user.signs")
-require("user.indent-blankline")
-vim.cmd([[hi NvimTreeNormal guibg=NONE ctermbg=NONE]])
-
+require("user.core")
+require("user")
