--- nvim/lua/.luarc.json
@@ -1,84 +1,0 @@
-{
-    "workspace.checkThirdParty": false,
-    "diagnostics.globals": [
-        "vim"
-    ],
-    "workspace.library": [
-        "/home/hemanth/.config/nvim",
-        "/home/hemanth/.local/share/nvim/site",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/Comment.nvim",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/LuaSnip",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/Navigator.nvim",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/better-escape.nvim",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/blue-moon",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/bufferline.nvim",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/catppuccin",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/cmp-buffer",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/cmp-cmdline",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/cmp-nvim-lsp",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/cmp-nvim-lua",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/cmp-path",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/cmp_luasnip",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/colorbuddy.vim",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/document-color.nvim",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/friendly-snippets",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/github-nvim-theme",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/gitsigns.nvim",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/gruber.vim",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/gruvbox.nvim",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/gruvbuddy.nvim",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/impatient.nvim",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/jellybeans.vim",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/kanagawa.nvim",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/lualine.nvim",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/lunar.nvim",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/material.nvim",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/neon",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/nlsp-settings.nvim",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/null-ls.nvim",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/nvim-autopairs",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/nvim-cheat.sh",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/nvim-cmp",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/nvim-colorizer.lua",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/nvim-dap",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/nvim-lsp-installer",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/nvim-lspconfig",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/nvim-rdark",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/nvim-solarized-lua",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/nvim-tmux-navigation",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/nvim-transparent",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/nvim-tree.lua",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/nvim-treesitter",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/nvim-treesitter-context",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/nvim-ts-context-commentstring",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/nvim-ts-rainbow",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/nvim-tundra",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/nvim-web-devicons",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/packer.nvim",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/playground",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/plenary.nvim",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/popfix",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/popup.nvim",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/telescope-file-browser.nvim",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/telescope-ui-select.nvim",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/telescope.nvim",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/toggleterm.nvim",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/tokyodark.nvim",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/tokyonight.nvim",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/vim-repeat",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/vim-surround",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/vscode.nvim",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/wal.vim",
-        "/tmp/.mount_nvimN9NSQ4/usr/share/nvim/runtime",
-        "/tmp/.mount_nvimN9NSQ4/usr/lib/nvim",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/catppuccin/after",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/cmp-buffer/after",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/cmp-cmdline/after",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/cmp-nvim-lsp/after",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/cmp-nvim-lua/after",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/cmp-path/after",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/cmp_luasnip/after",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/playground/after",
-        "/home/hemanth/.local/share/nvim/site/pack/packer/start/tokyodark.nvim/after"
-    ]
-}
