--- nvim/lua/user/colorizer.lua
@@ -1,2 +1,0 @@
--- Attaches to every FileType mode
-require 'colorizer'.setup()
