+++ nvim/lua/user/core/options.lua
