--- nvim/lua/user/firenvim-conf.lua
@@ -1,12 +1,0 @@
-vim.g.firenvim_config = {
-  globalSettings = { alt = "all" },
-  localSettings = {
-    [".*"] = {
-      cmdline = "neovim",
-      content = "text",
-      priority = 0,
-      selector = "textarea",
-      takeover = "never",
-    },
-  },
-}
