--- nvim/lua/user/indent-blankline.lua
@@ -1,9 +1,0 @@
-vim.opt.list = true
---[[ vim.opt.listchars:append "space:⋅" ]]
---[[ vim.opt.listchars:append "eol:↴" ]]
-
-require("indent_blankline").setup {
-    --[[ space_char_blankline = " ", ]]
-    show_current_context = true,
-    --[[ show_current_context_start = true, ]]
-}
