--- nvim/lua/user/lsp/settings/lua_ls.lua
@@ -1,16 +1,0 @@
-return {
-  settings = {
-
-    Lua = {
-      diagnostics = {
-        globals = { "vim" },
-      },
-      workspace = {
-        library = {
-          [vim.fn.expand("$VIMRUNTIME/lua")] = true,
-          [vim.fn.stdpath("config") .. "/lua"] = true,
-        },
-      },
-    },
-  },
-}
