--- nvim/lua/user/lsp/settings/pyright.lua
@@ -1,10 +1,0 @@
-return {
-  settings = {
-    python = {
-      analysis = {
-        typeCheckingMode = "off",
-        stubPath = "~/.config/typings",
-      },
-    }
-  },
-}
