--- nvim/lua/user/null-ls.lua
@@ -1,24 +1,0 @@
-local null_ls_status_ok, null_ls = pcall(require, "null-ls")
-if not null_ls_status_ok then
-  print("Something went wrong with null_ls")
-  return
-end
-
--- https://github.com/jose-elias-alvarez/null-ls.nvim/tree/main/lua/null-ls/builtins/formatting
-local formatting = null_ls.builtins.formatting
--- https://github.com/jose-elias-alvarez/null-ls.nvim/tree/main/lua/null-ls/builtins/diagnostics
--- local diagnostics = null_ls.builtins.diagnostics
-
--- https://github.com/prettier-solidity/prettier-plugin-solidity
--- npm install --save-dev prettier prettier-plugin-solidity
-null_ls.setup({
-  debug = false,
-  sources = {
-    formatting.prettierd.with({
-      extra_filetypes = { "toml", "solidity", "typescriptreact" },
-      --[[ extra_args = { "--single-quote", "--jsx-single-quote"}, ]]
-    }),
-    formatting.black.with({ extra_args = { "--fast" } }),
-    formatting.stylua,
-  },
-})
