+++ nvim/lua/user/plugins/colorschemes.lua
@@ -1,0 +1,30 @@
+return {
+  "rose-pine/neovim",
+  "tiagovla/tokyodark.nvim",
+  "marko-cerovac/material.nvim",
+  "folke/tokyonight.nvim",
+  "dylanaraps/wal.vim",
+  "ellisonleao/gruvbox.nvim",
+  "rebelot/kanagawa.nvim",
+  "nanotech/jellybeans.vim",
+  {
+    "catppuccin/nvim",
+    name = "catppuccin",
+  },
+  "projekt0n/github-nvim-theme",
+  "kyazdani42/blue-moon",
+  "tjdevries/colorbuddy.vim",
+  "RishabhRD/nvim-rdark",
+  "norcalli/nvim-colorizer.lua",
+  "lunarvim/lunar.nvim",
+  "lunarvim/darkplus.nvim",
+  "tjdevries/gruvbuddy.nvim",
+  "sam4llis/nvim-tundra",
+  "rafamadriz/neon",
+  "https://gitlab.com/madyanov/gruber.vim",
+  "ishan9299/nvim-solarized-lua",
+  "luisiacc/gruvbox-baby",
+  "ray-x/aurora",
+  "Shatur/neovim-ayu",
+  "bluz71/vim-nightfly-guicolors",
+}
