+++ nvim/lua/user/plugins/diffview.lua
@@ -1,0 +1,4 @@
+return {
+  "sindrets/diffview.nvim",
+  event = { "VimEnter" },
+}
