+++ nvim/lua/user/plugins/emmet.lua
@@ -1,0 +1,4 @@
+return {
+  "mattn/emmet-vim",
+  ft = { "html", "css", "markdown", "typescriptreact", "javascriptreact" },
+}
