+++ nvim/lua/user/plugins/init.lua
@@ -1,0 +1,22 @@
+return {
+  {
+    "alexghergh/nvim-tmux-navigation",
+    config = function()
+      local nvim_tmux_nav = require("nvim-tmux-navigation")
+
+      nvim_tmux_nav.setup({
+        disable_when_zoomed = true, -- defaults to false
+      })
+
+      vim.keymap.set("n", "<C-h>", nvim_tmux_nav.NvimTmuxNavigateLeft)
+      vim.keymap.set("n", "<C-j>", nvim_tmux_nav.NvimTmuxNavigateDown)
+      vim.keymap.set("n", "<C-k>", nvim_tmux_nav.NvimTmuxNavigateUp)
+      vim.keymap.set("n", "<C-l>", nvim_tmux_nav.NvimTmuxNavigateRight)
+    end,
+  },
+  {"RRethy/vim-illuminate", event = "BufReadPre"},
+  "nvim-lua/plenary.nvim", 
+  "tpope/vim-repeat",
+  "tpope/vim-surround",
+  "tpope/vim-fugitive",
+}
