+++ nvim/lua/user/plugins/lualine.lua
@@ -1,0 +1,39 @@
+return {
+  "nvim-lualine/lualine.nvim",
+  event = { "BufReadPre", "BufNewFile" },
+  config = function() 
+    local status_ok, lualine = pcall(require, "lualine")
+    if not status_ok then
+      return
+    end
+
+    lualine.setup {
+      options = {
+        icons_enabled = true,
+        theme = 'auto',
+        component_separators = { left = '', right = ''},
+        section_separators = { left = '', right = ''},
+        disabled_filetypes = {},
+        always_divide_middle = true,
+      },
+      sections = {
+        lualine_a = {'mode'},
+        lualine_b = {'branch','diff', 'diagnostics'},
+        lualine_c = {{'filename', path=1 }},
+        lualine_x = {'encoding', 'fileformat', 'filetype'},
+        lualine_y = {'progress'},
+        lualine_z = {'location'}
+      },
+      inactive_sections = {
+        lualine_a = {},
+        lualine_b = {},
+        lualine_c = {'filename'},
+        lualine_x = {'location'},
+        lualine_y = {},
+        lualine_z = {}
+      },
+      tabline = {},
+      extensions = {}
+    }
+  end
+}
