+++ nvim/lua/user/plugins/neogen.lua
@@ -1,0 +1,9 @@
+return {
+  "danymat/neogen",
+  event = { "BufReadPre", "BufNewFile" },
+  config = function()
+    require('neogen').setup {
+      snippet_engine = "luasnip"
+    }
+  end,
+}
