+++ nvim/lua/user/plugins/treesitter.lua
@@ -1,0 +1,90 @@
+return {
+  "nvim-treesitter/nvim-treesitter",
+  event = { "BufReadPre", "BufNewFile" },
+  build = ":TSUpdate",
+  dependencies = {
+    "windwp/nvim-ts-autotag",
+    "nvim-treesitter/nvim-treesitter-textobjects",
+  },
+  config = function()
+    -- import nvim-treesitter plugin
+    local treesitter = require("nvim-treesitter.configs")
+
+    -- configure treesitter
+    treesitter.setup({
+      ensure_installed = { "c", "javascript", "typescript", "python", "yaml", "markdown", "tsx" },
+      sync_install = false,
+      ignore_install = { "" },
+      autopairs = {
+        enable = true,
+      },
+      highlight = {
+        enable = true,
+        disable = { "" },
+        additional_vim_regex_highlighting = true,
+      },
+      indent = { enable = true, disable = { "yaml", "python" } },
+      context_commentstring = {
+        enable = true,
+        enable_autocmd = false,
+      },
+      autotag = {
+        enable = true,
+      },
+      incremental_selection = {
+        enable = true,
+        keymaps = {
+          init_selection = "<c-space>",
+          node_incremental = "<c-space>",
+          scope_incremental = "<c-s>",
+          node_decremental = "<M-space>",
+        },
+      },
+      textobjects = {
+        select = {
+          enable = true,
+          lookahead = true, -- Automatically jump forward to textobj, similar to targets.vim
+          keymaps = {
+            -- You can use the capture groups defined in textobjects.scm
+            ["aa"] = "@parameter.outer",
+            ["ia"] = "@parameter.inner",
+            ["af"] = "@function.outer",
+            ["if"] = "@function.inner",
+            ["ac"] = "@class.outer",
+            ["ic"] = "@class.inner",
+          },
+        },
+        move = {
+          enable = true,
+          set_jumps = true, -- whether to set jumps in the jumplist
+          goto_next_start = {
+            ["]m"] = "@function.outer",
+            --[']]'] = '@class.outer',
+          },
+          goto_next_end = {
+            ["]M"] = "@function.outer",
+            --[']['] = '@class.outer',
+          },
+          goto_previous_start = {
+            ["[m"] = "@function.outer",
+            --['[['] = '@class.outer',
+          },
+          goto_previous_end = {
+            ["[M"] = "@function.outer",
+            --['[]'] = '@class.outer',
+          },
+        },
+        swap = {
+          enable = true,
+          swap_next = {
+            ["<leader>a"] = "@parameter.inner",
+          },
+          swap_previous = {
+            ["<leader>A"] = "@parameter.inner",
+          },
+        },
+      },
+
+    })
+  end,
+}
